tools/bluemoon: Fix compilation error/warning
This patch fixes the following error/warning:

  CC tools/bluemoon.o
tools/bluemoon.c: In function ‘analyze_firmware’:
tools/bluemoon.c:813:10: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (len != le32_to_cpu(css->size) * 4) {
          ^
cc1: all warnings being treated as errors

+++ tools/bluemoon.c
@@ -810,7 +810,7 @@ static void analyze_firmware(const char *path)
 	printf("\n");
 
 
-	if (len != le32_to_cpu(css->size) * 4) {
+	if ((size_t) len != le32_to_cpu(css->size) * 4) {
 		fprintf(stderr, "CSS.size does not match file length\n");
 		goto done;
 	}
